* Move `mount.py` -> `base.py` * Move the mount creation helper function to a classmethod. * Move the simulator helper function to a classmethod on the the simulator mount.
