+++ src/panoptes/pocs/mount/constants.py
@@ -29,7 +29,8 @@ class SerialParams(Enum):
     MOVEMENT_STATUS = r'(?P<movement_status>[0-9])'
     TIME_SOURCE = r'(?P<time_source>[1-3])'
     HEMISPHERE = r'(?P<hemisphere>[01])'
-    FULL_STATUS = LONGITUDE + LATITUDE + GPS_STATUS + SYSTEM_STATUS + TRACKING_STATUS + MOVEMENT_STATUS + TIME_SOURCE + HEMISPHERE
+    FULL_STATUS = LONGITUDE + LATITUDE + GPS_STATUS + SYSTEM_STATUS + \
+        TRACKING_STATUS + MOVEMENT_STATUS + TIME_SOURCE + HEMISPHERE
 
 
 class GPS(IntEnum):
